Python_basic/requirements.txt
future==0.16.0
nltk==3.2.2
numpy==1.12.0
pbr==1.10.0
pyArango==1.2.8
python-arango==3.4.1
requests==2.12.4
six==1.10.0
stevedore==1.19.1
textblob==0.11.1
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
virtualenvwrapper-win==1.2.1
